+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Added `var_export` Twig extension.
+- Added a default list of `source_file_paths` for templates in order to support the
+  `var-export.php.twig` template.
+- Enabled Twig debug mode and added `DebugExtension` to allow for calling of `dump` in templates.
+
 ## [1.2.0] - 2021-05-20
 ### Added
 - Added sane defaults for `CodeDeploymentStateInterface` and `MaintenanceStrategyInterface` to allow